# api/openapi.yaml
openapi: 3.0.3
info:
  title: TechForge AI Agent Suite API
  description: |
    Enterprise-grade API for managing AI development agents
    
    The TechForge API provides programmatic access to deploy, manage, and orchestrate AI agents for software development tasks. 
    
    Features:
    - Deploy and manage 14 specialized AI agents
    - Execute tasks with intelligent orchestration
    - Real-time monitoring and analytics
    - Team coordination and collaboration
    - Enterprise security and compliance
    
    ## Authentication
    All requests require an API key in the Authorization header:
    ```
    Authorization: Bearer your-api-key
    ```
    
    ## Rate Limits
    - Starter: 100 requests/minute
    - Professional: 1000 requests/minute  
    - Enterprise: Unlimited
    
    ## WebSocket Support
    Real-time updates available via WebSocket at `wss://stream.techforge.ai/v2`
    
  version: 2.0.0
  contact:
    name: TechForge Support
    url: https://techforge.ai/support
    email: api-support@techforge.ai
  license:
    name: Proprietary
    url: https://techforge.ai/license
  termsOfService: https://techforge.ai/terms
servers:
  - url: https://api.techforge.ai/v2
    description: Production server
  - url: https://staging-api.techforge.ai/v2
    description: Staging server
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /agents/deploy:
    post:
      summary: Deploy an agent
      description: Deploy a specific AI agent for use
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployAgentRequest'
      responses:
        '201':
          description: Agent deployed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          $ref: '#/components/responses/PaymentRequired'
  /agents/{agentId}:
    get:
      summary: Get agent details
      tags: [Agents]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /tasks:
    get:
      summary: List tasks
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
  /tasks/{taskId}:
    get:
      summary: Get task details
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
        enum: [backend-expert, frontend-expert, database-expert, devops-expert, security-expert, qa-expert, ai-ml-expert, mobile-expert, game-expert, hardware-iot-expert, uiux-principal, web-design-agent, refactor-agent]
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
    Status:
      name: status
      in: query
      schema:
        type: string
        enum: [pending, in_progress, completed, failed, cancelled]
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        specialization:
          type: string
        tech_stack:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [available, deploying, active, busy, error, maintenance]
        pricing_tier:
          type: string
          enum: [starter, professional, enterprise]
        resource_requirements:
          type: object
          properties:
            cpu:
              type: number
            memory:
              type: string
            gpu:
              type: boolean
    DeployAgentRequest:
      type: object
      required: [agent_id]
      properties:
        agent_id:
          type: string
        config:
          type: object
        resource_limits:
          type: object
          properties:
            cpu:
              type: number
            memory:
              type: string
            timeout:
              type: integer
    DeploymentResponse:
      type: object
      properties:
        agent_id:
          type: string
        deployment_id:
          type: string
        status:
          type: string
        deployed_at:
          type: string
          format: date-time
        config:
          type: object
...
    TaskResult:
      type: object
      properties:
        task_id:
          type: string
        agent_id:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, failed, cancelled]
        output:
          oneOf:
            - type: string
            - type: object
        metrics:
          type: object
          properties:
            execution_time:
              type: number
            tokens_used:
              type: integer
            accuracy_score:
              type: number
        timestamp:
          type: string
          format: date-time
        error:
          type: string
    TaskList:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskResult'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
              code:
                type: integer
...
    PaymentRequired:
      description: Payment required - subscription limit reached
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
              upgrade_url:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
...